+++ paper.bib
@@ -1,5 +1,4 @@
 % Encoding: UTF-8
-
 @Article{barth:2014,
   author  = {Barth, A. and Beckers, J.-M. and Troupin, C. and Alvera-Azc\'arate, A. and Vandenbulcke, L.},
   journal = {Geoscientific Model Development},
@@ -20,7 +19,7 @@
   issn    = {1463-5003},
   pages   = {90-101},
   volume  = {52-53},
-  doi     = {https://doi.org/10.1016/j.ocemod.2012.05.002},
+  doi     = {10.1016/j.ocemod.2012.05.002},
   url     = {https://www.sciencedirect.com/science/article/pii/S1463500312000790},
 }
 
@@ -33,6 +32,7 @@
   editor    = {Fernando Loizides and Birgit Scmidt},
   pages     = {87--90},
   publisher = {IOS Press},
+  doi       = {10.3233/978-1-61499-649-1-87},
   url       = {https://eprints.soton.ac.uk/403913/},
 }
 
@@ -100,118 +100,7 @@
   address = {Exeter, Devon},
   year    = {2010 - 2015},
   url     = {https://scitools.org.uk/cartopy},
-}
-
-@Article{Shahzadi:2021,
-  author    = {Shahzadi, Kanwal and Pinardi, Nadia and Barth, Alexander and Troupin, Charles and Lyubartsev, Vladyslav and Simoncelli, Simona},
-  journal   = {Frontiers in Environmental Science},
-  title     = {A New Global Ocean Climatology},
-  year      = {2021},
-  issn      = {2296-665X},
-  month     = {Aug},
-  volume    = {9},
-  doi       = {10.3389/fenvs.2021.711363},
-  publisher = {Frontiers Media SA},
-  url       = {https://www.frontiersin.org/journals/environmental-science/articles/10.3389/fenvs.2021.711363/full},
-}
-
-@Article{Belgacem:2021,
-  author    = {Belgacem, Malek and Schroeder, Katrin and Barth, Alexander and Troupin, Charles and Pavoni, Bruno and Raimbault, Patrick and Garcia, Nicole and Borghini, Mireno and Chiggiato, Jacopo},
-  journal   = {Earth System Science Data},
-  title     = {{Climatological distribution of dissolved inorganic nutrients in the western Mediterranean Sea (1981–2017)}},
-  year      = {2021},
-  issn      = {1866-3516},
-  month     = {Dec},
-  number    = {12},
-  pages     = {5915–5949},
-  volume    = {13},
-  doi       = {10.5194/essd-13-5915-2021},
-  publisher = {Copernicus GmbH},
-  url       = {https://essd.copernicus.org/articles/13/5915/2021/},
-}
-
-@Article{Doglioni:2023,
-  author    = {Doglioni, Francesca and Ricker, Robert and Rabe, Benjamin and Barth, Alexander and Troupin, Charles and Kanzow, Torsten},
-  journal   = {Earth System Science Data},
-  title     = {Sea surface height anomaly and geostrophic current velocity from altimetry measurements over the Arctic Ocean (2011–2020)},
-  year      = {2023},
-  issn      = {1866-3516},
-  month     = jan,
-  number    = {1},
-  pages     = {225–263},
-  volume    = {15},
-  doi       = {10.5194/essd-15-225-2023},
-  publisher = {Copernicus GmbH},
-  url       = {https://essd.copernicus.org/articles/15/225/2023/},
-}
-
- 
-@Article{Webb:2025,
-  author    = {Webb, Thomas J. and Beja, Joana and Bejarano, Salvador Jesús Fernández and Ramos, Elvira and Sainz-Villegas, Samuel and Soetaert, Karline and Stolte, Willem and Troupin, Charles and Weigel, Benjamin},
[…]
