+++ paper.bib
@@ -103,4 +103,112 @@
   doi     = {10.5281/zenodo.1182735}
 }
 
+@Article{Belgacem:2021,
+  author    = {Belgacem, Malek and Schroeder, Katrin and Barth, Alexander and Troupin, Charles and Pavoni, Bruno and Raimbault, Patrick and Garcia, Nicole and Borghini, Mireno and Chiggiato, Jacopo},
+  title     = {{Climatological distribution of dissolved inorganic nutrients in the western Mediterranean Sea (1981–2017)}},
+  journal   = {Earth System Science Data},
+  year      = {2021},
+  volume    = {13},
+  number    = {12},
+  pages     = {5915–5949},
+  month     = {Dec},
+  issn      = {1866-3516},
+  doi       = {10.5194/essd-13-5915-2021},
+  publisher = {Copernicus GmbH},
+  url       = {http://dx.doi.org/10.5194/essd-13-5915-2021},
+}
+
+@Article{Shahzadi:2021,
+  author    = {Shahzadi, Kanwal and Pinardi, Nadia and Barth, Alexander and Troupin, Charles and Lyubartsev, Vladyslav and Simoncelli, Simona},
+  title     = {A New Global Ocean Climatology},
+  journal   = {Frontiers in Environmental Science},
+  year      = {2021},
+  volume    = {9},
+  month     = {Aug},
+  issn      = {2296-665X},
+  doi       = {10.3389/fenvs.2021.711363},
+  publisher = {Frontiers Media SA},
+  url       = {http://dx.doi.org/10.3389/fenvs.2021.711363},
+}
+
+@Article{Doglioni:2023,
+  author    = {Doglioni, Francesca and Ricker, Robert and Rabe, Benjamin and Barth, Alexander and Troupin, Charles and Kanzow, Torsten},
+  title     = {Sea surface height anomaly and geostrophic current velocity from altimetry measurements over the Arctic Ocean (2011–2020)},
+  journal   = {Earth System Science Data},
+  year      = {2023},
+  volume    = {15},
+  number    = {1},
+  pages     = {225–263},
+  month     = jan,
+  issn      = {1866-3516},
+  doi       = {10.5194/essd-15-225-2023},
+  publisher = {Copernicus GmbH},
+  url       = {http://dx.doi.org/10.5194/essd-15-225-2023},
+}
+
+@Article{Webb:2025,
+  author    = {Webb, Thomas J. and Beja, Joana and Bejarano, Salvador Jesús Fernández and Ramos, Elvira and Sainz-Villegas, Samuel and Soetaert, Karline and Stolte, Willem and Troupin, Charles and Weigel, Benjamin},
+  title     = {Realising the potential of interoperable data products to improve the outlook for marine biodiversity: Lessons from the European marine observation and data network},
+  journal   = {Marine Policy},
+  year      = {2025},
+  volume    = {173},
+  pages     = {106578},
+  month     = mar,
+  issn      = {0308-597X},
+  doi       = {10.1016/j.marpol.2024.106578},
+  publisher = {Elsevier BV},
+  url       = {http://dx.doi.org/10.1016/j.marpol.2024.106578},
+}
+
+@Article{Schaap:2010,
+  author    = {Schaap, Dick M.A. and Lowry, Roy K.},
+  title     = {SeaDataNet – Pan-European infrastructure for marine and ocean data management: unified access to distributed data sets},
+  journal   = {International Journal of Digital Earth},
+  year      = {2010},
+  volume    = {3},
+  number    = {sup1},
+  pages     = {50–69},
+  month     = apr,
+  issn      = {1753-8955},
[…]
