+++ README.md
@@ -86,7 +86,7 @@ For instance `DIVAnd` can used in projects such as:
 ### Binder
 
 Most notebooks need more resources that what is can currently available on Binder. The introduction notebooks (introduction to OI and variationa analysis) however work
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/gher-uliege/Diva-Workshops/master?filepath=notebooks%2F1-Intro%2F04-OI-variational-analysis-introduction.ipynb).
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/gher-uliege/Diva-Workshops/master?filepath=notebooks%2F1-Intro%2F1-04-OI-variational-analysis-introduction.ipynb).
 
 ## 🆘 Troubleshooting
 
