### 2025-01-31 6ac889b
CompatHelper: add new compat entry for GLMakie at version 0.11, (keep existing compat)
