+++ Project.toml
@@ -45,6 +45,7 @@ DIVAnd = "2"
 DataFrames = "1"
 DataStructures = "0.18"
 DelimitedFiles = "1"
+GLMakie = "0.11"
 GRIB = "0.4"
 GeoArrays = "0.9"
 GeoDatasets = "0.1"
