### 2025-02-03 270d722
remove GLMakie dependency
