+++ Project.toml
@@ -8,7 +8,6 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GRIB = "b16dfd50-4035-11e9-28d4-9dfe17e6779b"
 GeoArrays = "2fb1d81b-e6a0-5fc5-82e6-8e06903437ab"
 GeoDatasets = "ddc7317b-88db-5cb5-a849-8449e5df04f9"
