### 2025-04-02 #277 @arfon
> @openjournals/dev 🆘 We're having trouble adding a reviewer to this Pre-Review issue for JOSE. Help?

Looks like there was a bad deploy of Editorialbot. This should be fixed now.
