### 2025-04-02 #278 @editorialbot
[REVIEW]: DIVAnd training: producing climatologies with Jupyter notebooks
**Submitting author:** @ctroupin (<a href="http://orcid.org/0000-0002-0265-1021">Charles Troupin</a>)
**Repository:** https://github.com/gher-uliege/Diva-Workshops
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.1
