### 2025-04-02 #278 @kls2177
Hello @gaelforget and @Balinus,

Thank you again for agreeing to review. We are now ready to go!

Please complete the above checklists and then provide your specific review comments as issues to the [submitted repo](https://github.com/gher-uliege/Diva-Workshops).

If you have any questions about the process, please refer to the [JOSE docs](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html). If you don't find what you are looking for in these docs, feel free to connect with me.

Thanks!

Karen
