### 2025-05-16 27d9577
CompatHelper: bump compat for Interpolations to 0.16, (keep existing compat)
