+++ Project.toml
@@ -52,7 +52,7 @@ GeoMakie = "0.7"
 GeometryOps = "0.1"
 GeometryTypes = "0.8"
 IJulia = "1"
-Interpolations = "0.15"
+Interpolations = "0.15, 0.16"
 JupyterFormatter = "0.1"
 MAT = "0.10"
 Makie = "0.21, 0.22"
