### 2025-06-12 837ba6d
CompatHelper: bump compat for CairoMakie to 0.14, (keep existing compat)
