### 2025-06-19 gher-ulg/Diva-Workshops#67 @Balinus
Add Community guidelines
Review requirements ask for Community Guidelines (https://github.com/openjournals/jose-reviews/issues/278#issuecomment-2773156224).

"Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support"

Unless I'm mistaken, there is none currently.

Cheers!
