### 2025-06-22 a0c5951
CompatHelper: bump compat for Makie to 0.24, (keep existing compat)
