+++ Project.toml
@@ -55,7 +55,7 @@ IJulia = "1"
 Interpolations = "0.15, 0.16"
 JupyterFormatter = "0.1"
 MAT = "0.10"
-Makie = "0.21, 0.22, 0.23"
+Makie = "0.21, 0.22, 0.23, 0.24"
 Missings = "1"
 NCDatasets = "0.14"
 PhysOcean = "0.6"
