### 2025-06-30 a31d2e9
Update issue templates
