+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: bug
+assignees: ctroupin
+
+---
+
+## Describe the bug 
+...
+
+## Julia version
+```julia
+VERSION     
+```        
+
+## `DIVAnd` 
+```julia
+using Pkg
+Pkg.status("DIVAnd")        
+```
+
+## The name of the notebook that lead to the problem.      
+
+## Full screen output
+
+Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.        
+
+## Full stack trace with error message.
