### 2025-06-30 45a645e
Update issue templates
