+++ .github/ISSUE_TEMPLATE/contributing-to-the-module.md
@@ -1,0 +1,12 @@
+---
+name: Contributing to the module
+about: Propose a notebook describing a new application
+title: "[NEW]"
+labels: ''
+assignees: ctroupin
+
+---
+
+## Describe the objective of the new notebook
+
+## Explain why this notebook is a relevant addition to the repository
