### 2025-07-11 gher-uliege/Diva-Workshops#70 @Balinus
Activate env for each jupyter notebook
Currently, some notebooks does not activate the repo Project.toml

The following lines should be added at the beginning of each Notebooks:

```julia
import Pkg
Pkg.activate("/path/to/cloned/repo/Diva-Workshops")
Pkg.instantiate()
```

Related to https://github.com/openjournals/jose-reviews/issues/278
