### 2025-07-11 #278 @Balinus
Hello!

I have finished reviewing the paper/instruction module. This is very instructional and gives a lot of examples that I will personally reuse 😄

I have posted 2 (minor) issues in the repo:

https://github.com/gher-uliege/Diva-Workshops/issues/67
https://github.com/gher-uliege/Diva-Workshops/issues/70

Everything else seems quite good in my humble opinion and I was able to install the module's dependencies easily, navigate the examples and learn from it. I must admit I haven't tried to do the "homework" though 😆

So, as far as I'm concerned, this should be published once the minor issues are resolved.

Thanks!
