### 2025-07-22 gher-uliege/Diva-Workshops#70 @ctroupin
Thanks for the comment, the `activate` command has now been added to all the notebooks.
