### 2025-07-22 gher-uliege/Diva-Workshops#67 @ctroupin
We've added a file https://github.com/gher-uliege/Diva-Workshops/blob/main/CONTRIBUTING.md
and also the issues now come with templates which reflect the content of the `CONTRIBUTING.md` file.
