### 2025-07-22 #278 @ctroupin
Thanks for your encouraging comments and the in-depth review!

We would be obviously happy the notebooks are useful to you as well.
