### 2025-07-22 3e55e09
remove pygments installation
