### 2025-08-02 5e06d32
CompatHelper: bump compat for DataStructures to 0.19, (keep existing compat)
