+++ Project.toml
@@ -42,7 +42,7 @@ ColorSchemes = "3"
 Compat = "4"
 DIVAnd = "2"
 DataFrames = "1"
-DataStructures = "0.18"
+DataStructures = "0.18, 0.19"
 DelimitedFiles = "1"
 GRIB = "0.4"
 GeoArrays = "0.9"
