### 2025-08-13 99e8f91
Create CITATION.cff [skip ci]
