+++ CITATION.cff
@@ -1,0 +1,48 @@
+cff-version: 1.2.0
+title: Diva-Workshops
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Charles
+    family-names: Troupin
+    email: ctroupin@uliege.be
+    affiliation: 'GHER, University of Liège'
+    orcid: 'https://orcid.org/0000-0002-0265-1021' 
+  - given-names: Alexander
+    family-names: Barth
+    orcid: 'https://orcid.org/0000-0003-2952-5997'
+    email: A.Barth@uliege.be
+    affiliation: 'GHER, University of Liège'
+  - given-names: Jean-Marie
+    family-names: Beckers
+    email: JM.Beckers@uliege.be
+    affiliation: 'GHER, University of Liège'
+    orcid: 'https://orcid.org/0000-0002-7045-2470'
+  - given-names: Sylvain
+    family-names: Watelet
+    affiliation: 'Royal Meteorological Institute of Belgium'
+    orcid: 'https://orcid.org/0000-0002-1080-3988'
+  - given-names: 'Katrin'
+    family-names: Leinweber
+    affiliation: 'Technische Informationsbibliothek (TIB)'
+    orcid: 'https://orcid.org/0000-0001-5135-5758'
+identifiers:
+  - type: doi
+    value: 10.5281/zenodo.1248029
+    description: Zenodo
+  - type: swh
+    value: 'swh:1:dir:073695a641642aefcde6fda9becf8ebc8d708337' 
+repository-code: 'https://github.com/gher-uliege/Diva-Workshops'
+keywords:
+  - Oceanography
+  - Data interpolation
+  - Data analysis
+  - Julia
+  - Notebooks
+  - Training
+license: MIT
+commit: 796d30018a5c8e1c26a9714c8e67fae0a7f77cef
+version: v1.3.1
+date-released: '2025-01-28'
