### 2025-09-15 #278 @kls2177
Hello @natgeo-wong, Thank you for agreeing to review.

Please use the following command to generate your individual checklist:

`@editorialbot generate my checklist`

And please provide your specific review comments as issues to the [submitted repo](https://github.com/gher-uliege/Diva-Workshops). I will work to generate a new checklist for you.

If you have any questions about the process, please refer to the [JOSE docs](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html). If you don't find what you are looking for in these docs, feel free to connect with me.

Thanks!

Karen
