### 2025-09-16 #278 @gaelforget
> Version: Does the release version given match the repository release?

Version indicated above is `1.3.1`. Current repository release is `1.3.2`.  Seems satisfying.

More importantly, the repo does not contain `Manifest.toml`. This means that the versions of various packages  to be used are not fully constrained (incl. `DIVANd.jl` or `NCDatasets.jl`). Shouldn't one be added to increase reproducibility in the future?

As an example of why this matters is that `3-09-full-analysis.ipynb` and other notebooks currently fail in my environment, seemingly in relation to an issue with `NCDatasets.jl` version -> https://github.com/gher-uliege/Diva-Workshops/issues/72
