### 2025-09-16 gher-uliege/Diva-Workshops#72 @gaelforget
[BUG]
## Describe the bug

`NCDatasets` (`v0.14.8`) errors out trying to read file.

## Julia version
```julia
VERSION
```

`v1.11.6`

## `DIVAnd`
```julia
using Pkg
Pkg.status("DIVAnd")
```

`v2.7.12`

## The name of the notebook that lead to the problem.

`3-09-full-analysis.ipynb`

## Full screen output

Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.
