### 2025-09-16 gher-uliege/Diva-Workshops#73 @gaelforget
Installation instructions: Is there a clearly stated list of dependencies?
The `Project.toml` contains a list packages that includes a `compat` section to constrain dependency versions, which is used in the notebooks via `Pkg.activate("../..")` in the notebooks. But none of this is obvious when one first lands in the repo.

Could this be documented in the `README.md`?

( linking https://github.com/openjournals/jose-reviews/issues/278#issuecomment-3297789408 )
