### 2025-09-16 gher-ulg/Diva-Workshops#74 @gaelforget
Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?
top level `README.md` seems more focused on the `DIVAnd.jl` package than on the workshop material. Confusing to me.

It would be good to highlight the notebook folder more clearly in the `README.md`, and for example :

- provide a list of subfolders + short description
- mention that the notebooks will get dependencies via Project.toml
- mention the secondary `notebooks/README.md` (maybe other READMEs I may have overlooked?)

The details in `About DIVAnd` could be moved to the package doc, with a link here is needed. Same for the `install Julia` part, which could be moved to a separate file.
