### 2025-09-16 #278 @gaelforget
wrt the `JOSE paper` part of the checklist :

> Description: Does the paper describe the learning materials and sequence?

To make things more obvious I'd suggest turning the `The notebooks have been organised by sub-folders according to their objectives: ` paragraph into an itemized list in the top level README.md :

*update : the above suggestion has been completed*

```
1. Introduction ...
2. Preprocessing ...
...
```

( https://github.com/gher-uliege/Diva-Workshops/issues/74#issuecomment-3298686042 )

> Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

Could you add something about how many sessions are suggested, and how much time per session, to go through the material?

Something like `1. Session 1 : Intro notebooks, 4 hours` etc
