### 2025-09-16 #278 @gaelforget
wrt `Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)`

Would be good to add a more explicit pointer to the paper in the README -- saying that `Instructional design` and other aspects are covered in more detail in JOSS paper.

> Learning objectives: Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)

In the paper yes, but would be good to repeat this in the `README.md`. Might draw new potential new users in, helping them decide to go read the paper for more information.

> Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)

Not sure how much expertise I have on this, but would report that

- I felt a bit overwhelmed with the sheer number of notebooks -- having the sense I needed to go through them in sequence, and no obvious option other than using Jupyter to run notebooks, I found myself manually launching every notebook one after the other, and clicking a few times for each notebook.
- would have liked to be pointed to a select few essential notebook(s) ; e.g. for a quick start see `3-07-example-analysis.ipynb` and a couple others.
- some of this probably has to do more with my own user preferences, but I still don't know of a good way browse quickly through jupyter notebook collection folders (VS code?) -- a list of links in markdown that point to nb files on GitHub (e.g. [3-07-example-analysis.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/3-Analysis/3-07-example-analysis.ipynb)) would help
