### 2025-09-16 gher-uliege/Diva-Workshops#74 @gaelforget
> `README.md` seems more focused on the `DIVAnd.jl` package than on the workshop material. Confusing.
>
> It would be good to highlight the notebook folder more in the `README.md`, and for example :
>
> * provide a list of subfolders + short description
> * mention that the notebooks will get dependencies via Project.toml
>
> The details in `About DIVAnd` could be moved to the package doc, with a link here.

I just realized there was [another README](notebooks/README.md) I had overlooked. Suggest adding cross reference in the top level README to this one.
