+++ contributing.md
@@ -1,0 +1,39 @@
+# Contributing to this project
+
+Many of the notebooks presented in this repository were created following requests by users!     
+We welcome suggestions and improvements from them!                
+This document helps you find the best way to make this project better.
+
+We primarily communicate through GitHub issues to ensure transparency and collaboration. If this is not working for you, please feel free to contact the project maintainers via email.
+
+## Code of Conduct
+
+We are committed to fostering an open, inclusive, and respectful community for all contributors to the `Diva-Workshops` project. 
+
+We ask every contributor to follow the guidelines: 
+- Treat all participants with respect, kindness, and professionalism, regardless of their background, experience, or perspectives. 
+- Engage in constructive and collaborative communication, avoiding any form of harassment, discrimination, or offensive behavior. 
+- Provide feedback thoughtfully and accept it gracefully. 
+
+These principles have to be applied in all interactions: code, documentation, issues, and pull requests. 
+
+## Got a Question?
+
+If the question relates to 
+- the notebooks from this repos: feel free to [open a new issue](https://github.com/gher-uliege/Diva-Workshops/issues/new?template=question.md).
+- to `DIVAnd`: check the [corresponding documentation](https://gher-uliege.github.io/DIVAnd.jl/stable/) and if needed, [submit an issue](https://github.com/gher-uliege/DIVAnd.jl/issues/new/choose).
+- to the Julia language: ask your question on the [Discourse channel](https://discourse.julialang.org/).
+
+## Found a Bug?
+
+If there is something wrong with a notebook, help us and [submit an issue](https://github.com/gher-uliege/Diva-Workshops/issues/new?template=bug_report.md).
+
+## Missing a Feature?
+
+We've done our best to guide you through the process of creating a climatology, from data ingestion to generating the final product.      
+If you find that:
+- a notebook lacks clear explanations or needs improvement,
+- a new notebook is needed to address a specific part of the workflow,
+- a notebooks should be added to support additional data types (other than oceanography),
+
+please share your suggestions by opening a [new issue](https://github.com/gher-uliege/Diva-Workshops/issues/new?template=contributing-to-the-module.md). 
