+++ CONTRIBUTING.md
@@ -1,29 +1,49 @@
 # Notes for Contributors
 
-The developments of [`DIVAnd`](https://github.com/gher-uliege/DIVAnd.jl/) and the present set of notebooks are lost of the time user-driven:        
-if you consider something is missing and should be improved, don't hesitate to make your voice heard!      
-First you need to know whether your chance is related to
-1. `DIVAnd`: need of new functionalities, improved documentation, bug...
-2. The notebooks: please follow the instructions below.
+The developments of [`DIVAnd`](https://github.com/gher-uliege/DIVAnd.jl/) and the present set of notebooks are most of the time user-driven:        
+if you consider something is missing and should be improved, don't hesitate to make your voice heard!    
 
-## Improving `DIVAnd`
+This document helps you find the best way to make this project better.
 
-[Submit an issue](https://github.com/gher-uliege/DIVAnd.jl/issues/new/choose) or contact the main developers by mail.      
+## Code of Conduct
 
-## Improving the notebooks
+We are committed to fostering an open, inclusive, and respectful community for all contributors to the `Diva-Workshops` project. 
 
-### Contributing to the module
+We ask every contributor to follow the guidelines: 
+- Treat all participants with respect, kindness, and professionalism, regardless of their background, experience, or perspectives. 
+- Engage in constructive and collaborative communication, avoiding any form of harassment, discrimination, or offensive behavior. 
+- Provide feedback thoughtfully and accept it gracefully. 
 
-If you have a notebook showing a `DIVAnd` application that doesn't appear in the present set of notebooks, or would like to have an exemple notebook providing a workflow not described in other notebooks, you can either:
-1. Submit a pull-request.
-2. Open a [new issue](https://github.com/gher-uliege/Diva-Workshops/issues/new/choose) describing your application.
-3. Contact the developers by mail, if you don't have a GitHub account.
+These principles have to be applied in all interactions: code, documentation, issues, and pull requests. 
+
+## Seeking support
+ 
+The notebooks are supposed to be self-explanatory, but if some parts are not clear enough, indicate the part of the notebook that has to be improved.
+
+When help is needed with a specific application, it is always good to have access to the original data (or a subset of it) and the notebook that lead to the problem.
+
+## Got a Question?
+
+If the question relates to 
+- the notebooks from this repos: feel free to [open a new issue](https://github.com/gher-uliege/Diva-Workshops/issues/new?template=question.md).
+- to `DIVAnd`: check the [corresponding documentation](https://gher-uliege.github.io/DIVAnd.jl/stable/) and if needed, [submit an issue](https://github.com/gher-uliege/DIVAnd.jl/issues/new/choose).
+- to the Julia language: ask your question on the [Discourse channel](https://discourse.julialang.org/).     
+
+## Missing a feature?
+
+We've done our best to guide you through the process of creating a climatology, from data ingestion to generating the final product.      
+If you find that:
+- a notebook lacks clear explanations or needs improvement,
+- a new notebook is needed to address a specific part of the workflow,
+- a notebooks should be added to support additional data types (other than oceanography),
+
+please share your suggestions by opening a [new issue](https://github.com/gher-uliege/Diva-Workshops/issues/new?template=contributing-to-the-module.md). 
 
 Ideally, the datasets used in the notebooks should be publicly available.
 
-### Reporting issues or problems with the repository 
+### Found a bug? 
 
-The preferred way to report an issue is also through the GitHub Issue.           
+The preferred way to report an issue is also through the [GitHub Issue]((https://github.com/gher-uliege/Diva-Workshops/issues/new?template=contributing-to-the-module.md)).           
[…]
