### 2025-09-16 gher-uliege/Diva-Workshops#72 @gaelforget
> Thanks! I believe this one is related to [`PhysOcean`](https://github.com/gher-uliege/PhysOcean.jl) and this issue: [gher-uliege/PhysOcean.jl#17](https://github.com/gher-uliege/PhysOcean.jl/issues/17).
>
> So we need to use the master branch of `PhysOcean` instead of the last release.

OK. This works at present time. Once new release is out, it would be good to specify that version in `Manifest.toml` & `Project.toml/compat`.
