### 2025-09-16 gher-uliege/Diva-Workshops#72 @ctroupin
Yes, thanks, we'll make sure that the new version of PhysOcean is in the files.
