### 2025-09-16 c568140
remove readme from notebooks dir
