### 2025-09-16 gher-uliege/Diva-Workshops#75 @gaelforget
[BUG] PyPlot needed for `4-09-open-boundary-conditions.ipynb`
## Describe the bug
...

`PyPlot` needed for `4-09-open-boundary-conditions.ipynb` is missing from `Project.toml`

## Julia version
```julia
VERSION
```

## `DIVAnd`
```julia
using Pkg
Pkg.status("DIVAnd")
```

## Full screen output

Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.
