### 2025-09-16 gher-ulg/Diva-Workshops#76 @gaelforget
[BUG] missing file for `4-13-BlackSea-detrend-geostrophic.ipynb`
## Describe the bug

`4-13-BlackSea-detrend-geostrophic.ipynb` errors out due to lack of a file.

## Julia version
```julia
VERSION
```

## `DIVAnd`
```julia
using Pkg
Pkg.status("DIVAnd")
```

## Full screen output

Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.
