### 2025-09-16 #278 @gaelforget
## running the notebooks

I finished running all the notebooks (1- to 4- series + advanced topics).

To summarize :

- Most ran to completion without any issue.  Nice!
- Five errored out ( https://github.com/gher-uliege/Diva-Workshops/issues/78, https://github.com/gher-uliege/Diva-Workshops/issues/77, https://github.com/gher-uliege/Diva-Workshops/issues/76 , https://github.com/gher-uliege/Diva-Workshops/issues/75 , https://github.com/gher-uliege/Diva-Workshops/issues/72 )
- Three took a long time that may warrant warning to users (~ 10' maybe , `4-04-heatmapsCV.ipynb`, `4-05-turtles-tensity-map.ipynb`, and `4-09-open-boundary-conditions.ipynb`)

## notebooks content

Impressive work! The notebook collection is comprehensive and didactic. Went over it quickly but this will be very useful and has been. Powerful and mature tool for a an important task in oceanography.
