### 2025-09-16 gher-ulg/Diva-Workshops#77 @gaelforget
[BUG] calling `mean` makes notebook error out
## Describe the bug

cannot run `1-02-Julia-introduction.ipynb` notebook to completion using e.g. `utils/run-all-notebooks`

because notebook error out on `mean`. Possible fix : replace with `?mean` to demonstrate that the method is not found.

## Julia version
```julia
VERSION
```

## `DIVAnd`
```julia
using Pkg
Pkg.status("DIVAnd")
```

## Full screen output

Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.
