### 2025-09-16 gher-ulg/Diva-Workshops#78 @gaelforget
[BUG] notebook `2-03-data-downloading.ipynb` errors out
## Describe the bug

- `import Pkg` does not export `PackageSpec` so `Pkg.add(PackageSpec(name="PhysOcean", rev="master"))` errors out
- quick "fix" could be to expand to `Pkg. PackageSpec` , but hacking the environment from within one notebook this way seems ill-advised , especially when pointing to a master branch in whatever state it is at the time the user call this

## Julia version
```julia
VERSION
```

## `DIVAnd`
```julia
using Pkg
Pkg.status("DIVAnd")
```

## Full screen output

Preferably obtained after setting `ENV["JULIA_DEBUG"] = "DIVAnd"`.
