### 2025-09-17 #278 @gaelforget
Have read the paper. Looks great.

Adding a paragraph on _how someone might adopt it_ seems needed based on the checklist. Maybe just saying something like this would work?

_download the GitHub repo, follow instructions from the readme, the different notebook folders can be covered independently (?), plan to devote an average of 10 minutes (?) per notebook in class._
