### 2025-09-17 gher-uliege/Diva-Workshops#75 @ctroupin
@Alexander-Barth do you agree if I use `Makie` (instead of `PyPlot`) in some of the examples (I'd prefer to avoid `PyPlot` in the notebooks because we had frequently issues when participants installed it, due to different Python versions etc).
