### 2025-09-19 #278 @ctroupin
Thanks a lot @gaelforget , I've solved most of the issues and addressed the comments, I'll carefully reply to the issues to the changes made are clear.
