### 2025-09-19 gher-uliege/Diva-Workshops#75 @ctroupin
This is now completed: the example scripts in `DIVAnd` that are called by `4-09-open-boundary-conditions.ipynb` are now working with `Makie`.
