+++ Project.toml
@@ -58,7 +58,6 @@ MAT = "0.10"
 Makie = "0.21, 0.22, 0.23, 0.24"
 Missings = "1"
 NCDatasets = "0.14"
-PhysOcean = "0.6"
 Roots = "2"
 SpecialFunctions = "2"
 Statistics = "1"
