### 2025-09-19 89e57fb
clean and update README: merged the other README (with the description of the subfolder) with the main README (more clarity for users) added a sentence to refer to the [CONTRIBUTING.md](CONTRIBUTING.md) file moved content more related to DIVAnd to the corresponding repository added a note concerning the dependencies and the `Project.toml` file
