+++ README.md
@@ -10,19 +10,23 @@
 
 ## Objective
 
-The [Jupyter](https://jupyter.org/) notebooks contained in this repository are designed to explain users how to create gridded fields from in situ observations using the [`DIVAnd`](https://github.com/gher-uliege/divand.jl) software tool.
+The [Jupyter](https://jupyter.org/) notebooks contained in this repository help users to create gridded fields from in situ observations using the [`DIVAnd`](https://github.com/gher-uliege/divand.jl) software tool.     
+More information about `DIVAnd` is available in the package [documentation](https://gher-uliege.github.io/DIVAnd.jl/stable/).
 
-## Content
+### Content
 
-The notebooks are organised into 5 categories, separated in folders:
-* [1-Intro](./notebooks/1-Intro/): know the basics commands in Julia, read/write netCDF files and create different types of plots (scatter, histograms, maps, ...).
-* [2-Preprocessing](./notebooks/2-Preprocessing): learn about the input file preparation: observations (download and reading), bathymetry and mask, time periods, ...
-* [3-Analysis](./notebooks/3-Analysis/): perform different types of analysis, optimise the analysis parameters and work with different coordinate systems.
-* [4-AdvancedTopics](./notebooks/4-AdvancedTopics/): discover more complex types of analysis, using for instance advection or inequality constraints.
+The notebooks are organised into 4 categories, according to their main objectives.
 
-## How to use the notebooks?
+| Folder | Content | Recommended duration |
+|:-------------------|:----------------------:|----------------|
+[1-Intro](./notebooks/1-Intro/) | Know the basics commands in Julia, read/write netCDF files and create different types of plots (scatter, histograms, maps, ...) |  |
+[2-Preprocessing](./notebooks/2-Preprocessing) | Learn about the input file preparation: observations (download and reading), bathymetry and mask, time periods, ... |  |
+[3-Analysis](./notebooks/3-Analysis/) | Perform different types of analysis, optimise the analysis parameters and work with different coordinate systems |   |
+[4-AdvancedTopics](./notebooks/4-AdvancedTopics/) | Discover more complex types of analysis, using for instance advection or inequality constraints. |   |
 
-Different paths can be taken, depending on your experience.
+### How to use the notebooks?
+
+Different paths can be taken, depending on your experience with Julia and with `DIVAnd`.
 
 | Experience in Julia | Experience with DIVAnd | Recommendation |
 |:-------------------:|:----------------------:|----------------|
@@ -36,88 +40,43 @@ Different paths can be taken, depending on your experience.
 
 ### Julia 
 
-Julia language can be installed using `juliaup`, as detailed in the "Download" section of the Julia web: https://julialang.org/downloads/.     
-A Julia session is started by typing `julia` in the terminal or by clicking on the Julia shortcut in Windows.
-
-```bash
-$ julia
-               _
-   _       _ _(_)_     |  Documentation: https://docs.julialang.org
-  (_)     | (_) (_)    |
-   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
-  | | | | | | |/ _` |  |
-  | | |_| | | | (_| |  |  Version 1.11.3 (2025-01-21)
- _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
-|__/                   |
-julia> 
-```
+Julia language can be installed using `juliaup`, as detailed in the [Download](https://julialang.org/downloads/) section of Julia.
 
 ### Jupyter
 
-Jupyter has to be installed in order to have a notebook interface.        
[…]
