+++ README.md
@@ -18,7 +18,7 @@ More information about `DIVAnd` is available in the package [documentation](http
 
 ### Content
 
-The notebooks are organised into 4 categories, according to their main objectives and following the "Instructional design" described in the paper.    
+The notebooks are organised into 4 categories, according to their main objectives and following the "Instructional design" described in the associated paper in JOSE.    
 The different notebook folders can be covered independently, depending on the experise of the users.       
 Suggestions for the number of sessions and time to be dedicated for each category are provide below.
 
@@ -29,6 +29,56 @@ Suggestions for the number of sessions and time to be dedicated for each categor
 [3-Analysis](./notebooks/3-Analysis/) | Perform different types of analysis, optimise the analysis parameters and work with different coordinate systems | 2 sessions of 4 hours   |
 [4-AdvancedTopics](./notebooks/4-AdvancedTopics/) | Discover more complex types of analysis, using for instance advection or inequality constraints. | 4 hours  |
 
+<details>
+
+<summary><h3>Full list of notebooks</h3></summary>
+
+[1-01-notebooks-basics.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/1-Intro/1-01-notebooks-basics.ipynb)         
+[1-02-Julia-introduction.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/1-Intro/1-02-Julia-introduction.ipynb)          
+[1-03-netCDF.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/1-Intro/1-03-netCDF.ipynb)           
+[1-04-OI-variational-analysis-introduction.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/1-Intro/1-04-OI-variational-analysis-introduction.ipynb)        
+[1-05-plots-maps.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/1-Intro/1-05-plots-maps.ipynb)           
+[2-01-topography.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/2-Preprocessing/2-01-topography.ipynb)              
+[2-02-reading-data.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/2-Preprocessing/2-02-reading-data.ipynb)           
+[2-03-data-downloading.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/2-Preprocessing/2-03-data-downloading.ipynb)           
+[2-04-ODV-data-import.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/2-Preprocessing/2-04-ODV-data-import.ipynb)              
+[2-05-defining-time-periods.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/2-Preprocessing/2-05-defining-time-periods.ipynb)         
+[3-01-L-and-epsilon-effect.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/3-Analysis/3-01-L-and-epsilon-effect.ipynb)          
+[3-02-correlation-length.ipynb](https://github.com/gher-uliege/Diva-Workshops/blob/main/notebooks/3-Analysis/3-02-correlation-length.ipynb)          
[…]
