### 2025-09-19 #278 @ctroupin
Here are the replies to your comments and suggestions. Thanks for this in-depth review and testing of the notebooks!

> Version indicated above is 1.3.1. Current repository release is 1.3.2. Seems satisfying.

- [x] Apologies for that: we improved the 1.3.1 version following the comments from the 1st reviewer and made a release afterwards.

> More importantly, the repo does not contain Manifest.toml. This means that the versions of various packages to be used are not fully constrained (incl. DIVANd.jl or NCDatasets.jl). Shouldn't one be added to increase reproducibility in the future?

- [x] The [Manifest.toml](https://github.com/gher-uliege/Diva-Workshops/blob/main/Manifest.toml file has now been added to the project.

> Could you add something about how many sessions are suggested, and how much time per session, to go through the material?

- [x] We created a new table in the README where we explain the content of the folders and suggest the number of sessions and the time needed for them (even if that depends on many factors).

> Would be good to add a more explicit pointer to the paper in the README -- saying that Instructional design and other aspects are covered in more detail in JOSS paper.

- [x] This was also added to the README

> In the paper yes, but would be good to repeat this in the README.md. Might draw new potential new users in, helping them decide to go read the paper for more information.

- [x] Also added to the README (i.e., repeating the objectives).

> I felt a bit overwhelmed with the sheer number of notebooks -- having the sense I needed to go through them in sequence, and no obvious option other than using Jupyter to run notebooks, I found myself manually launching every notebook one after the other, and clicking a few times for each notebook.

and

> would have liked to be pointed to a select few essential notebook(s) ; e.g. for a quick start see 3-07-example-analysis.ipynb and a couple others.

- [x] Addressed in the section [How to use or adopt the notebooks?](https://github.com/gher-uliege/Diva-Workshops?tab=readme-ov-file#how-to-use-or-adopt-the-notebooks) and also through a [Fast start](https://github.com/gher-uliege/Diva-Workshops/tree/main?tab=readme-ov-file#-fast-start) paragraph

> some of this probably has to do more with my own user preferences, but I still don't know of a good way browse quickly through jupyter notebook collection folders (VS code?) -- a list of links in markdown that point to nb files on GitHub (e.g. 3-07-example-analysis.ipynb) would help.

- [x] We've added the [list of notebooks](https://github.com/gher-uliege/Diva-Workshops/tree/main?tab=readme-ov-file#full-list-of-notebooks) with links

> Five errored out ( [BUG] notebook 2-03-data-downloading.ipynb errors out gher-uliege/Diva-Workshops#78, [BUG] calling mean makes notebook error out gher-uliege/Diva-Workshops#77, [BUG] missing file for 4-13-BlackSea-detrend-geostrophic.ipynb gher-uliege/Diva-Workshops#76 , [BUG] PyPlot needed for 4-09-open-boundary-conditions.ipynb gher-uliege/Diva-Workshops#75 , [BUG] gher-uliege/Diva-Workshops#72 )

- [x] The issues were solved.

> Three took a long time that may warrant warning to users (~ 10' maybe , 4-04-heatmapsCV.ipynb, 4-05-turtles-tensity-map.ipynb, and 4-09-open-boundary-conditions.ipynb)

- [x] We've added a short text to warn the users that it will take time.

> Adding a paragraph on how someone might adopt it seems needed based on the checklist. Maybe just saying something like this would work?
> download the GitHub repo, follow instructions from the readme, the different notebook folders can be covered independently (?), plan to devote an average of 10 minutes (?) per notebook in class.

- [x] Done on the README, where we indicate an approximate time to spend on each subset of notebook.
