### 2025-09-19 #278 @kls2177
Thanks @gaelforget! @ctroupin, I have added a reviewer, @natgeo-wong, so it would be great to get one more independent review based on your recent updates before moving forward.

Once we hear from @natgeo-wong, I will outline the next steps.
