### 2026-01-06 #278 @kls2177
Hello all, a new year's resolution - let's get this paper to the finish line! @natgeo-wong thanks for starting your review. Do you have specific comments to share with @ctroupin to help them revise and allow you to check off the unchecked items?
