### 2026-01-12 #278 @natgeo-wong
[_Review__DIVAnd_Training__Producing_Climatologies_with_Jupyter_Notebooks.pdf](https://github.com/user-attachments/files/24573221/_Review__DIVAnd_Training__Producing_Climatologies_with_Jupyter_Notebooks.pdf)

It's been a bit, but I did get some comments up. So sorry for the delay, and hope my comments are useful.

Thanks!
