### 2026-02-05 #278 @ctroupin
Dear @natgeo-wong natgeo-wong,

thanks for your review and advice, and please allow me to reply point-by-point hereinafter:

> I understand the ubiquity of Jupyter notebooks in this day and age. However, one comment I’d like
to suggest for the future is the usage of Pluto.jl notebooks, as these notebooks are more reactive
and allow for real-time modification of input and automatic visualization of the output. For a
module especially focused on the usage of DIVAnd it is always better in my opinion to have reactive
notebooks like Pluto.jl or marimo (for python) than Jupyter notebooks. Nonetheless, the choice
has been made (for now) and I respect it.

We totally agree; we have been using `Pluto.jl` in other projects and it clearly provides improvements with respect to Jupyter.
However in our experience with Pluto, there was often a rather long delay at the beginning of the notebook. In addition in deployments on the cloud, we encountered frequent "Kernel restart" issues.
This is why, for our training/teaching activities, we preferred to stick to Jupyter.

Meanwhile, we added a new notebook (in Pluto): `3-09-full-analysis.jl`, which has the same content of the Jupyter notebook of the same name. The `README` has been updated accordingly.

> One more package to suggest: the authors of this project often refer to source code, and activate
projects that are in the parent folder, necessitating relative paths. I would suggest the usage of
DrWatson.jl and the @quickactivate macro to minimize the errors that could potentially result
from relative paths should notebooks change directories in the future.

The suggestion was taken into account and `DrWatson.jl` is now used in the notebooks, in particular the commands
- `@quickactivate`
- `datadir()` and
- `plotsdir()`.

> **Documentation**
On Community Guidelines, while the author understands that DIVAnd is mostly maintained
and worked upon by GHER, I don’t really see any documentation or guidelines for contribution to
the module. However, this is a minor issue, given that I foresee that most of the word and usage of
DIVAnd will come from GHER.

While it is true that most of the developments and maintenance is performed by the GHER staff, we welcome contributions from all users. We already had a [CONTRIBUTING.md](https://github.com/gher-uliege/Diva-Workshops/blob/review_JOSE_3/CONTRIBUTING.md) document, referenced within the main `README` file.

> **Pedagogy / Instructional Design**
The author believes that the pedagogical design of the module makes it relatively easy to follow,
in the sense that we know what each module is supposed to do. However, as someone who prefers
breaking up the code into multiple parts, sometimes it might be better to give labels to show what
each of the different figures are doing. I have noted that, for example, in notebook 3-13-polar-
coordinates, there are a list of figure codes produced at the bottom but no mention of what exactly
those figures are for. If some of those figures are not needed, would be good to remove them from
the notebooks to prevent confusion.

We acknowledge that in some notebooks are lacking information and labels. Our approach was to make those notebooks available in the repository, even if their maturity (i.e. documentation, labeling etc) was not as good as the others.
We modify `3-13-polar-coordinates` according to the comment and will continue to enrich those notebooks which may lack explanation.

> Other than that, the module seems to be comprehensive in both the basic and advanced topics,
and I look forward to seeing how this module can be used among the Julia Earth Sciences community
in the future.

Thanks! We also hope that this module will contribute to have more `DIVAnd` users.

__Note:__ the changes related to this review were added to the branch [review_JOSE_3](https://github.com/gher-uliege/Diva-Workshops/tree/review_JOSE_3), so they are easier to identify.
