### 2026-03-17 #278 @ctroupin
Dear @kls2177 @natgeo-wong ,

I would like to add a project (who provided funding) in the acknowledgements, is it OK?
Also, have you seen the [last reply](https://github.com/openjournals/jose-reviews/issues/278#issuecomment-3855047818) ?

Thanks

Charles
