### 2026-04-07 #278 @ctroupin
Thanks @natgeo-wong .
@kls2177 I've (finally) added the missing acknowledgment.
All the changes are in [review_JOSE_3](https://github.com/gher-uliege/Diva-Workshops/tree/review_JOSE_3), shall I merge with the main branch now?
With regards
