### 2026-04-14 #278 @kls2177
Hi @ctroupin, I made a few minor suggestions to the paper.md document. The major change I suggested is to the title, adding the word "oceanographic". Let me know if you feel this is reasonable. I'm an atmospheric scientist, so I felt that this needed to be added based on the applications in the notebooks. Once you approve or reject my suggestions, I will begin the publication process.

Please also make a Zenodo archive and report the DOI in the thread.

Thanks!
