### 2026-04-17 #278 @ctroupin
Not yet. I'm checking here: https://github.com/gher-uliege/Diva-Workshops/blob/main/paper.md
(I did include the work "ocean" following your suggestion so that the title now reads

> DIVAnd training: producing ocean climatologies with Jupyter notebooks

)
