### 2026-05-06 #278 @kls2177
Hello @ctroupin, for the Zenodo archive, the title and the list of authors must be the same as the JOSE paper. It looks like you have more authors listed for the archive and the title is not the same as the paper. Can you please update this?

Also, is this the correct DOI for Gandin (1965): [https://doi.org/10.1002/qj.49709239320](https://doi.org/10.1002/qj.49709239320)? If so, can you please add this?

Thanks!
