### 2026-05-06 #278 @ctroupin
Dear @kls2177 ,

thanks, I've made the suggested changes:
- The DOI has been modified
- The Zenodo record has been amended in order to have the same title and authors.
