### 2026-05-06 #278 @kls2177
Hi @ctroupin, here are a few final notes on the paper (based on the pdf generated above):

Line 6: “creation” -> “the creation”
Line 7: remove “the”
Line 16: remove “the”
Line 23: “method” -> “methods”
Line 23: should this Bretherton citation be in parentheses?
Line 27: I think you need to connect this new paragraph with the previous sentence. Something like: “DIVA (Data-Interpolating Variational Analaysis; Troupin et al., 2012) is comparable to optimal interpolation, but it takes coastlines, sub-basins and advection into account. DIVA is an analysis method that is based on …” *I copied most of this from the DIVA github*
Line 40: the formatting of the numbered list is not working.
Line 81: the formatting of the numbered list is not working.
References: The DOI for Gandin (1965) is still missing in the pdf that I generated.
